requirements.txt
pyrez~=1.1.1.1
python-dotenv~=1.0.0
pymongo~=4.4.0
requests~=2.28.2
py-cord~=2.4.1